Optimisation pass (#93)
- Only call FastLED updates when necessary
- Omit certain RGB update functions if respective LED counts are 0
- Only update Rainbow lighting effects when necessary
- Add warning for high BEEF_LED_REFRESH values
- Optimise ticker, move some static class variables to global scope to reduce initialisation guards
- Fix RGB animation normalisation factors being wrong
- Reduce code size by removing parameters for globals
- Omit entire LED update function if both ring and light bar LEDs are set to 0

+++ fw/analog_button.h
@@ -9,7 +9,7 @@ typedef struct {
   // Number of ticks we need to advance before recognizing an input
   uint8_t deadzone;
   // How long to sustain the input before clearing it (if opposite direction is input, we'll release immediately)
-  uint32_t sustain_ms;
+  uint8_t sustain_ms;
   // Always provide a zero-input for one poll before reversing?
   bool clear;
 
@@ -26,7 +26,7 @@ typedef struct {
 extern analog_button button_x;
 extern analog_button button_y;
 
-void analog_button_init(analog_button* self, uint8_t deadzone, uint32_t sustain_ms, bool clear);
+void analog_button_init(analog_button* self, uint8_t deadzone, uint8_t sustain_ms, bool clear);
 int8_t analog_button_poll(analog_button* self, uint32_t current_value);
 #ifdef __cplusplus
 }
